requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.35.87
botocore==1.35.87
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
defusedxml==0.7.1
Flask==3.0.2
Flask-Cors==4.0.0
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
lxml==5.3.0
MarkupSafe==2.1.5
pillow==11.0.0
pydantic==2.10.4
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-pptx==1.0.2
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.0.1
XlsxWriter==3.2.0
youtube-transcript-api==0.6.3
